==> analysis/11-scenic-analysis.Rmd <==
---
title: "SCENIC analysis"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

In this document we are going to analyse the results from the pySCENIC command-line software. SCENIC is a tool for single-cell regulatory network inference and clustering.

## Setup

Define chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/11-scenic-analysis.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE,
  fig.align = "center",
  out.width = '100%'
)
```

Load required packages:

```{r}
library(pheatmap)
library(scater)
library(scran)
library(viridis)
library(hues)
```

Import experiment object:

```{r}
sce <- readRDS("data/09-cell-annotation.rds")
```

Import activity matrix:

```{r}
auc <- read.csv("data/pyscenic-activity.csv", row.names = "Regulon", stringsAsFactors = FALSE)

auc <- as.matrix(auc)
```

Setup celltype colour palette:

```{r}
celltype_factor <- factor(sce$celltype)

celltype_levels <- levels(celltype_factor)

celltype_colours <- MouseGastrulationData::EmbryoCelltypeColours[celltype_levels]
```

Setup cluster colour palette:

```{r}
cluster_factor <- factor(sce$cluster)

cluster_levels <- levels(cluster_factor)

cluster_colours <- iwanthue(length(cluster_levels))

names(cluster_colours) <- cluster_levels
```

Setup heatmap annotations and palette:

```{r}
ann <- data.frame(Celltype = sce$celltype, Cluster = sce$cluster, row.names = colnames(sce))

col <- list(Celltype = celltype_colours, Cluster = cluster_colours)

key <- seq(0, 1, length.out = 101)
```

## Activity

Plot matrix of AUC values for all regulons:

```{r fig.width = 10, fig.height = 10}
pheatmap(
  mat = auc,
  color = viridis(100),
  breaks = key,
  annotation_col = ann,
  annotation_colors = col,
  border_color = NA,
  show_rownames = FALSE,
  show_colnames = FALSE
)
```

Plot matrix of AUC values for regulons with the highest average activity:

```{r fig.width = 10, fig.height = 10}
ave <- rowMeans(auc)

idx <- which(ave >= sort(ave, decreasing = TRUE)[50], arr.ind = TRUE)

pheatmap(
  mat = auc[idx, ],
  color = viridis(100),
  breaks = key,
  annotation_col = ann,
  annotation_colors = col,
  border_color = NA,
  show_rownames = TRUE,
  show_colnames = FALSE
)
```

Plot matrix of AUC values for regulons with the highest variation:

```{r fig.width = 10, fig.height = 10}
var <- rowVars(auc)

idx <- which(var >= sort(var, decreasing = TRUE)[50], arr.ind = TRUE)

pheatmap(
  mat = auc[idx, ],
  color = viridis(100),
  breaks = key,
  annotation_col = ann,
  annotation_colors = col,
  border_color = NA,
  show_rownames = TRUE,
  show_colnames = FALSE
)
```
